move conversion to own module as a feature
